\name{emm}
\alias{emm}
\docType{data}
\title{
	Data object representing the EMM model
}
\description{
	Data object produced when the GenModel function is executed.
}
\details{
	Data object produced when GenModel function is executed. This data object can be used for plotting and classification purposes. 
}
\seealso{
\code{\link{createGenDB}}
}

\keyword{datasets}
