\name{MMSA}
\Rdversion{1.1}
\alias{MMSA}
\docType{package}
\title{
Build a classification hierarchy from fasta files which can be queried and used to build EMM models.
MMSA
}
\description{
More about what it does (maybe more than one line)
This package reads fasta formatted files and builds a classification hierarchy from it.
It also stores the sequences and classification in a SQLite databases, which can be queried easily.
The package can be used to build EMM models which are later used to classify unknown sequences. }
\details{
\tabular{ll}{
Package: \tab MMSA\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2009-10-12\cr
License: \tab What license is it under?\cr
LazyLoad: \tab yes\cr
}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
Who wrote it

Maintainer: Who to complain to <yourfault@somewhere.net>
~~ The author and/or maintainer of the package ~~
}
\references{
~~ Literature or other references for background information ~~
}
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
